Fix a bug about parameter ordering
We've mistakenly assigned Wind to Water and Water to Wind at
the beginning of the program. We didn't realize it till this
point because we were also sending our parameters in wrong order.
So 2 mistakes together made the error invisible :) This is a good
point to change all the parameters to have the same order.